Update fhir-server-test #1053
- Remove unused LOINC.java
- Refactor the testng.xml (package,classes) usage
- Fix $apply operation issue where detail was not set
- Move tests with special considerations to their own packages
	- bulkdata, db2, kafka
	- introduced controls in test.properties to disable or enable tests
- Enabled all Tests
