--- fhir-server-test/src/test/java/com/ibm/fhir/server/test/LOINC.java
@@ -1,22 +1,0 @@
-/*
- * (C) Copyright IBM Corp. 2019
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package com.ibm.fhir.server.test;
-
-public class LOINC {
-    public static final String URI = "http://loinc.org";
-    public static final String OID = "urn:oid:2.16.840.1.113883.6.1";
-    public static final String CCDA_CCD = "34133-9";
-    
-    public static final String CCDA_SECTION_ALLERGIES = "48765-2";  
-    public static final String CCDA_SECTION_MEDICATIONS = "10160-0";
-    public static final String CCDA_SECTION_PROBLEMLIST = "11450-4";    
-    public static final String CCDA_SECTION_PROCEDURES = "47519-4";
-    public static final String CCDA_SECTION_RESULTS = "30954-2";
-    public static final String CCDA_SECTION_SOCIALHISTORY = "29762-2";
-    public static final String CCDA_SECTION_VITALSIGNS = "8716-3";
-    public static final String CCDA_SECTION_ENCOUNTERS = "46240-8";
-}
