+++ fhir-server-test/src/test/resources/test.properties
@@ -34,3 +34,12 @@ fhirclient.logging.enabled = false
 fhirclient.hostnameVerification.enabled = false
 fhirclient.http.receive.timeout = 60000
 fhirclient.http.return.pref = minimal
+
+# Used to turn on or off tests or throttle
+test.client.oauth.enabled = false
+test.websocket.enabled = false
+test.kafka.enabled = false
+test.bulkdata.import.enabled = false
+test.bulkdata.export.enabled = false
+test.db2.enabled = false
+test.performance.default = 1
